Fix for https://github.com/rse/grunt-traceur-simple/issues/2
Fix for https://github.com/rse/grunt-traceur-simple/issues/2 (escaping
parentheses prevents plugin from finding the Node executable on Windows
64-bit machines). I also found it useful to print the command that
failed, because it was only by running it myself and fiddling around
with the path that I worked out what the problem was.

I don't know if the other special characters need to be escaped or not,
possibly not. I left them alone for now.